// src/shared/api/types/ExerciseSet.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Measurement } from './Measurement.ts'

export const exerciseSetSetTypeEnum = {
	normal: 'normal',
	warmup: 'warmup',
	dropset: 'dropset',
	myorep: 'myorep',
	failure: 'failure',
} as const

export type ExerciseSetSetTypeEnumKey =
	(typeof exerciseSetSetTypeEnum)[keyof typeof exerciseSetSetTypeEnum]

export type ExerciseSet = {
	/**
	 * @description The number of the set within the exercise.
	 * @type integer
	 */
	setNumber: number
	/**
	 * @description The type of the set, such as normal, warmup, dropset, myorep, or failure.
	 * @type string
	 */
	setType?: ExerciseSetSetTypeEnumKey | null
	/**
	 * @description Indicates whether this set was completed.
	 * @type boolean | undefined
	 */
	setCompleted?: boolean
	/**
	 * @description When true, marks this set as the start of a new linked set group (drop set or myo reps) even if preceded by another set of the same type.
	 * @type boolean
	 */
	isDropsetStart?: boolean | null
	/**
	 * @description List of measurement codes and their units used for this exercise, in order
	 * @type array
	 */
	setMeasurements: Measurement[]
}