// src/shared/api/types/WorkoutData.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ExerciseAdditionalData } from './ExerciseAdditionalData.ts'
import type { ExerciseData } from './ExerciseData.ts'
import type { ExerciseGroupType } from './ExerciseGroupType.ts'
import type { MeasurementTemplate } from './MeasurementTemplate.ts'
import type { WorkoutDataType } from './WorkoutDataType.ts'

export type WorkoutData = {
	/**
	 * @description Unique identifier for the exercise.
	 * @type string | undefined
	 */
	exerciseCode?: string
	/**
	 * @description Name of the exercise.
	 * @type string | undefined
	 */
	exerciseName?: string
	/**
	 * @description Additional notes or instructions for this specific instance of the exercise.
	 * @type string | undefined
	 */
	exerciseNotes?: string
	/**
	 * @type array | undefined
	 */
	measurementTemplate?: MeasurementTemplate[]
	/**
	 * @description URL to a thumbnail image for this exercise.
	 * @type string
	 */
	exerciseThumbnailUrl?: string | null
	/**
	 * @description The profile ID of creator whos thumbnail / video will be prioritized
	 * @type string | undefined, uuid
	 */
	creatorProfileId?: string
	/**
	 * @description Indicates if the exercise is performed with weight on each side.
	 * @type boolean
	 */
	isWeightPerSide?: boolean | null
	/**
	 * @description Denormalized ExerciseTypeCode of the exercise (e.g. WEIGHT_REPS, VIDEO_ONLY). Lets clients apply type-specific behavior (like VIDEO_ONLY\'s locked duration) without resolving the exercise.
	 * @type string
	 */
	exerciseTypeCode?: string | null
	/**
	 * @description Array of alternate exercise codes that can be used for this exercise.
	 * @type array | undefined
	 */
	alternateExerciseCodes?: string[]
	/**
	 * @description The tempo for the exercise, if applicable. Format is typically \"X-X-X\" (e.g., \"2-0-1\").
	 * @type string
	 */
	exerciseTempo?: string | null
	/**
	 * @description The order of this exercise within the workout or its group.
	 * @type integer | undefined
	 */
	exerciseOrder?: number
	/**
	 * @description Identifier for the group this exercise belongs to, if any. Null if not part of a group.
	 * @type integer
	 */
	exerciseGroupId?: number | null
	/**
	 * @description Represents the type of an exercise group in a workout.\n
	 * @type string | undefined
	 */
	exerciseGroupType?: ExerciseGroupType | null
	/**
	 * @description If exerciseGroupType is SECTION then we can set a custom name here.
	 * @type string
	 */
	exerciseGroupName?: string | null
	/**
	 * @description Represents a special type of workout data entry. Null for regular exercises.\n
	 * @type string | undefined
	 */
	workoutDataType?: WorkoutDataType | null
	/**
	 * @description Additional data for special workout data types. Each WorkoutDataType uses a subset of these properties.
	 * @type object | undefined
	 */
	additionalData?: ExerciseAdditionalData | null
	/**
	 * @type array | undefined
	 */
	exerciseData?: ExerciseData
}